refactor(typescript): Migration to TypeScript
All the code is in TypeScript but it does not leverage types yet.
Type declarations are rolled up by API Extractor but are not exposed in package.json yet.
JSDoc reads the JS code generated from Typescript.
Jest reports type errors as part of the test run; minor changes in the code had to be made to allow it.
ESLint is still in use but it does not run on any file.

re #24